+++ src/components/SiderMenu/BaseMenu.js
@@ -3,8 +3,8 @@ import { Menu, Icon } from 'antd';
 import Link from 'umi/link';
 import isEqual from 'lodash/isEqual';
 import memoizeOne from 'memoize-one';
-import pathToRegexp from 'path-to-regexp';
 import { urlToList } from '../_utils/pathTools';
+import { getMenuMatches } from './SiderMenuUtils';
 import styles from './index.less';
 
 const { SubMenu } = Menu;
@@ -22,14 +22,6 @@ const getIcon = icon => {
   }
   return icon;
 };
-
-export const getMenuMatches = (flatMenuKeys, path) =>
-  flatMenuKeys.filter(item => {
-    if (item) {
-      return pathToRegexp(item).test(path);
-    }
-    return false;
-  });
 
 export default class BaseMenu extends PureComponent {
   constructor(props) {
