+++ src/e2e/home.e2e.js
@@ -7,7 +7,7 @@ describe('Homepage', () => {
   it('it should have logo text', async () => {
     await page.goto(BASE_URL);
     await page.waitForSelector('h1', {
-      timeout: 2000,
+      timeout: 5000,
     });
     const text = await page.evaluate(() => document.getElementsByTagName('h1')[0].innerText);
     expect(text).toContain('Ant Design Pro');
